Added functionality to display the confirmed order after placing it.
in app.py (burgerorderer) i added logic to store the confirmed order in a variable (confirmed_order) after the user clicks "Place Order." This ensures the final order is passed to the template for display. The code is not done yet, there is still important codes in html to add.