Adding html for the functionality "confimed order"
My groupmember added functionality in app.py for confirmed orders to show a message, now i completed the functionality by adding code in index.html, a new section that checks if confirmed_order exists. If it does, it displays the confirmed burger and customizations at the bottom of the page, letting the user know their order has been successfully placed.