Modified the functionality of the code so it shows what you ordered in the message
I added to the existing code to ensure that the order of what the user has ordered are displayed in the confirmation message.

I started by ensuring that the order details could be stored in the session. When the user clicks the "Place Order" button, I added code that collects all information from the cart and stores it in a variable called final_order. Additionally, I created a variable named order_confirmed, which stores the confirmed order details so that they can be displayed on the page later.