Updated Dockerfile so the Database start to finally work..
Before, the app did not want to fetch the data (menu) from the database, so what i did is changed the code in the Dockerfile. The purpose of the change was that the original command to start the application was CMD ["python", "app.py"]. This would only run the app.py file without executing the MenuStore_db.py script first.
By changing the command to CMD ["sh", "-c", "python MenuStore_db.py && python app.py"], i ensured that the MenuStore_db.py script runs before the Flask application starts. This script is responsible for populating the MongoDB database with burger and customization data. Running MenuStore_db.py initializes the database with necessary entries (burgers and customizations), which allows the Flask application to fetch and display this data when it starts.
This ensures that both hamburgers and customizations are available to be rendered on the application’s front end.